% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinant.R
\name{determinant,WoodburyMatrix,logical-method}
\alias{determinant,WoodburyMatrix,logical-method}
\title{Calculate the determinant of a matrix}
\usage{
\S4method{determinant}{WoodburyMatrix,logical}(x, logarithm)
}
\arguments{
\item{x}{A object that is a subclass of \code{WoodburyMatrix}}

\item{logarithm}{Logical indicating whether to return the logarithm of the
matrix.}
}
\value{
Same as \link[base]{determinant}.
}
\description{
Calculates the (log) determinant of a \code{\linkS4class{WoodburyMatrix}}
using the matrix determinant lemma.
}
\seealso{
\link[base]{determinant}
}
