% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiate.R
\name{instantiate}
\alias{instantiate}
\alias{instantiate,GWoodburyMatrix-method}
\alias{instantiate,SWoodburyMatrix-method}
\title{Instantiate a matrix}
\usage{
instantiate(x)

\S4method{instantiate}{GWoodburyMatrix}(x)

\S4method{instantiate}{SWoodburyMatrix}(x)
}
\arguments{
\item{x}{Implicit matrix to directly instantiate.}
}
\value{
The directly instantiated matrix.
}
\description{
This is a generic to represent direct instantiation of an implicitly defined
matrix. In general this is a bad idea, because it removes the whole purpose
of using an implicit representation.
}
\section{Functions}{
\itemize{
\item \code{instantiate(GWoodburyMatrix)}: Method for general matrices.

\item \code{instantiate(SWoodburyMatrix)}: Method for symmetric matrices.

}}
\seealso{
\link{WoodburyMatrix}, \linkS4class{WoodburyMatrix}
}
