% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WoodburyMatrix.R
\name{WoodburyMatrix}
\alias{WoodburyMatrix}
\title{Create a Woodbury matrix identity matrix}
\usage{
WoodburyMatrix(
  A,
  B,
  U = Diagonal(nrow(B)),
  V = Diagonal(ncol(B)),
  X = Diagonal(nrow(B)),
  O,
  symmetric
)
}
\arguments{
\item{A}{Matrix \eqn{A} in the definition above.}

\item{B}{Matrix \eqn{B} in the definition above.}

\item{U}{Matrix \eqn{U} in the definition above. Defaults to a diagonal matrix.}

\item{V}{Matrix \eqn{V} in the definition above. Defaults to a diagonal matrix.}

\item{X}{Matrix \eqn{X} in the definition above. Defaults to a diagonal matrix.}

\item{O}{Optional, precomputed value of \eqn{O}, as defined above. THIS IS
NOT CHECKED FOR CORRECTNESS, and this argument is only provided for advanced
users who have precomputed the matrix for other purposes.}

\item{symmetric}{Logical value, whether to create a symmetric or general
matrix. See Details section for more information.}
}
\value{
A \code{\linkS4class{GWoodburyMatrix}} object for a non-symmetric
matrix, \code{\linkS4class{SWoodburyMatrix}} for a symmetric matrix.
}
\description{
Creates an implicitly defined matrix representing the equation
\deqn{A^{-1} + U B^{-1} V,}
where \eqn{A, U, B} and \eqn{V} are n x n, n x p, p x p and p x n matrices,
respectively. A symmetric special case is also possible with
\deqn{A^{-1} + X B^{-1} X',}
where \eqn{X} is n x p and \eqn{A} and \eqn{B} are additionally symmetric.
The available methods are described in \link{WoodburyMatrix-class}
and in \link{solve}.
}
\details{
The benefit of using an implicit representation is that the inverse of this
matrix can be efficiently calculated via
\deqn{A - A U O^{-1} V A}
where \eqn{O = B + VAU}, and its determinant by
\deqn{det(O) det(A)^{-1} det(B)^{-1}.}
These relationships are often called the Woodbury matrix identity and the
matrix determinant lemma, respectively. If \eqn{A} and \eqn{B} are sparse or
otherwise easy to deal with, and/or when \eqn{p < n}, manipulating the
matrices via these relationships rather than forming $W$ directly can have
huge advantageous because it avoids having to create the (typically dense)
matrix \deqn{A^{-1} + U B^{-1} V} directly.

Where applicable, it's worth using the symmetric form of the matrix. This
takes advantage of the symmetry where possible to speed up operations, takes
less memory, and sometimes has numerical benefits. This function will create
the symmetric form in the following circumstances:
\itemize{
  \item \code{symmetry = TRUE}; or
  \item the argument \code{X} is provided; or
  \item \code{A} and \code{B} are symmetric (according to
  \code{\link{isSymmetric}}) and the arguments \code{U} and \code{V} are
  NOT provided.
}
}
\seealso{
\linkS4class{WoodburyMatrix}, \link{solve}, \link{instantiate}
}
